src/main/java/com/example/todoapp/service/TodoService.java
package com.example.todoapp.service;

import com.example.todoapp.model.Todo;
import com.example.todoapp.repository.TodoRepository;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class TodoService {

    private final TodoRepository todoRepository;

    public TodoService(TodoRepository todoRepository) {
        this.todoRepository = todoRepository;
    }

    public List<Todo> getAll(){
        return this.todoRepository.findAll();
    }

    public Todo addTodo(Todo newTodo){
        return this.todoRepository.save(newTodo);
    }
}
